backend/requirements.txt

fastapi
uvicorn
python-multipart
python-dotenv
langchain
langchain-google-genai
langgraph
pydantic
slowapi
tenacity
google-generativeai
langchain-groq
langchain-openai
qdrant-client
pandas
plotly
pytest
httpx
